requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.97.0
google-adk==1.3.0
pydantic==2.11.4
cloudpickle==3.1.1
google-api-python-client==2.172.0
google-genai==1.20.0
python-dotenv
